function fix(ResultView methods's return type changed, for unit test)
