+++ src/main/java/codesquad/ladder/LadderMain.java
@@ -1,5 +1,8 @@
 package codesquad.ladder;
 
+import codesquad.ladder.controller.LadderController;
+import codesquad.ladder.view.ResultView;
+
 public class LadderMain {
 
     public static void main(String[] args) {
