+++ src/main/java/codesquad/ladder/model/exceptions/InvalidNumPeopleException.java
@@ -1,4 +1,4 @@
-package codesquad.ladder.exceptions;
+package codesquad.ladder.model.exceptions;
 
 public class InvalidNumPeopleException extends Exception {
 }
