+++ src/main/java/codesquad/ladder/model/exceptions/InvalidSizeLadderException.java
@@ -1,4 +1,4 @@
-package codesquad.ladder.exceptions;
+package codesquad.ladder.model.exceptions;
 
 public class InvalidSizeLadderException extends Exception{
 }
