+++ src/main/java/codesquad/ladder/view/InputView.java
@@ -1,4 +1,4 @@
-package codesquad.ladder;
+package codesquad.ladder.view;
 
 import java.util.*;
