--- src/test/java/codesquad/ladder/InputViewTest.java
@@ -1,12 +1,0 @@
-package codesquad.ladder;
-
-import static org.junit.Assert.*;
-
-import org.junit.Test;
-
-import java.util.ArrayList;
-
-public class InputViewTest {
-
-}
-
