--- src/test/java/codesquad/ladder/LadderControllerTest.java
@@ -1,8 +1,0 @@
-package codesquad.ladder;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-
-public class LadderControllerTest {
-
-}
